Update mbs to Spring Boot 2
Changes to pom.xml are all done to get the correct versions of
dependencies. Code changes are very minor.

+++ server/mbservice/src/main/java/org/rliz/mbs/recording/boundary/RecordingBoundaryService.java
@@ -73,7 +73,7 @@ public class RecordingBoundaryService {
         if (pageable.getOffset() >= recordings.size()) {
             return new PageImpl<>(new ArrayList<>(), pageable, recordings.size());
         }
-        int first = pageable.getOffset();
+        int first = (int) pageable.getOffset();
         int last = Math.min(first + pageable.getPageSize(), recordings.size());
         return new PageImpl<Recording>(recordings.subList(first, last), pageable, recordings.size());
     }
